backend/requirements.txt
# FastAPI and ASGI Server
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Database and ORM
sqlalchemy==2.0.30
psycopg2-binary==2.9.9

# Geospatial Libraries
geoalchemy2==0.14.5
shapely==2.0.4

# Data Validation and Settings
pydantic==2.7.1
python-dotenv==1.0.1

# API Utilities
slowapi==0.1.9

# Data Handling and Geocoding
pandas==2.2.2
requests==2.32.3
pgeocode==0.4.0